Ships can become desynced by tring to the same spot while completely still https://jira.ccpgames.com/browse/TQ-130155
Jamies theory (see defect comments) was pretty much spot on; SetBallPosition on the server would be sent to both clients at tick N, but the 'other' client didn't have the ball in its park so the action was meaningless there. The server then evolved park, generating velocities for both balls.

At tick N+1, each client would be notified of the 'other' ball only, so it would remain oblivious to the fact that itself had gathered momentum. In theory, using the balls old velocity rectifies the momentum bit, but as we don't send state updates for a ball to its own client, the client was also unaware of itself having moved in the previous tick, so that approach only changed the amount and direction of the desync.

So, proposed fix here is to exclude items from being collision candidates on the server during the tick they get moved between bubbles. That should guarantee that all parties in the bubble start simulating a given ball in the same tick.

Review: http://codecollaborator/ui#review:id=17177

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/eve/src/": change = 1440598]