Ships can become desynced by tring to the same spot while completely still https://jira.ccpgames.com/browse/TQ-130155
A better approach dawned on me this morning as I was fixing a edge case in the previous approach; instead of adding SetBallPosition to the system history, first call directly into destiny_SetBallPosition.  This way the moved ball will be at the desired position when we updates the bubbles at the start of the tick. We then proceed to process the history normally, including synchronizing the position of the ball on all clients to the frame boundary. This can potentially lead to a small jitter for observers when the ball is moving when AddBalls is called, as the rendering will extrapolate its position from that point until we execute the SetBallPosition a bit (less than a second) later.

Review: http://codecollaborator/ui#review:id=17180

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/eve/src/": change = 1440974]