Cloaking ships not removed correctly while in structure outside view and someone in control https://jira.ccpgames.com/browse/TQ-131334
When considering cloak/uncloak events, it's not enought to just look at the bubbleInteractives, we have to look at the clientInterests too when distributing Remove-/AddBall updates to the bubble.

Also changed bubbleInteractives from being a dict of dicts to be a dict of sets as thats what it's being used as, apart from being slightly faster and more memory efficient.

Review: http://codecollaborator/ui#review:id=17236

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/eve/src/": change = 1449995]