[eve MAINLINE] Automatic VCXPROJ submission from 1213
This might create build loops, but it was deemed valuable enough to be able to sync these.
https://teamcity.evetech.net/viewLog.html?buildId=4965664
!nobuild

Source Changes
----------------
[Eve Engine] 1458279 by jamieb (Jamie Bannister)
De-Sync in Abyssal Deadspace can be caused by Triglavian gate
https://jira.ccpgames.com/browse/TQ-129935
Follow-up to 1458208 - The DestinyTest project is treating forward declarations differently to the main Destiny project. Hopefully this works, but I can't test it locally because DestinyTest doesn't work out of the perforce-box
----------------
[Eve Engine] 1458249 by jamieb (Jamie Bannister)
De-Sync in Abyssal Deadspace can be caused by Triglavian gate
https://jira.ccpgames.com/browse/TQ-129935
Follow-up to 1458208 - The "destinyTest" project also includes a reference to the now-deleted "Hashers.h" file that needs cleaning up
----------------
[Eve Engine] 1458208 by jamieb (Jamie Bannister)
De-Sync in Abyssal Deadspace can be caused by Triglavian gate
https://jira.ccpgames.com/browse/TQ-129935
Short explanation:
This desync is caused by the ordering of ballpark items not matching between server and clients, leading to collisions having differing responses. The fix is to change the way we store these items so that server and clients will always sort them in the same way.
Long explanation follows:
These issues all seem to come down to data structures that are not sorted consistently between server and clients. Specifically:
* The collection of child boxes in a parent box
* The collection of balls in a box
* The collection of static-collidables in a box
* The collection of boxes that a ball/collidable is in
These structures are all supposed to maintain insertation order for the contained elements. But this has two flaws:
* They don't actually maintain insertion order. The first 8 elements are ordered as expected, but when the 9th element is inserted the order is lost. (Internally it looks like the unsorted_set maintains a tree with 8 buckets. Then when it gets 9 elements it grows to 64 buckets and rebalances the tree, seemingly losing the previous ordering)
* Insertion order isn't a viable strategy, as clients will not see items added/removed to space in the same order as the server (or even other clients) depending on how they move around.
(See changelists 95589 and 875165 for previous significant changes in this area)
To fix this I'm changing to std::set containers, with custom sorting functions:
* SortedSetOfBoxes - Sorts based on X index, then Y index, then Z index, then level
* SortedSetOfBalls - Sorts based on ball ID (for regular balls), or based on parentBall ID then ball ID (for miniballs/staticCollidables)
* SortedSetOfStaticCollidables - Sorts based on parentBall ID then ball ID (for miniballs/staticCollidables)
StaticCollidables now track which parent ball they belong to (mParentBallID). Miniballs already did this (mOwnerId)
I've also inclu...
----------------

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/eve/src/": change = 1458302]