Destiny desync caused by following a cloaking ship in the last moment, part 2 https://jira.ccpgames.com/browse/TQ-131389
The gist of this change is that now, instead of making the server transform CloakBall into RemoveBall for all clients except the cloaky one, the server just forwards CloakBall to the bubble and the clients therein remove the ball if it isn't their ball.  That allows us to explicitly stop all followers as the cloak action is applied on both/all endpoints.

After taking the ApplyDamageStates bit of the pre-tick apart, it became apparent that it didn't really do much except the cloak/uncloak shennanigans which can now be performed 'inline' in the main event loop, as opposed to collecting the events and then repeating a loot of busy looping to process proper events and such.

The risky bit here (imo) is that I've deleted the mechanism whereby we could potentially have multiple characterIDs associated with a single cloaked shipID.  Until structures are allowed to cloak, I can't imagine a use case for it, but this is very old code so meh.

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/eve/src/": change = 1478001]