Anti Bump Tethering Proposal - Tech Spike https://jira.ccpgames.com/browse/TFO-3055
This adds a new mode for a destiny ball; DSTBALL_MOORED and is mostly implemented in accordance to 
https://wiki.ccpgames.com/display/TF0/Design+Concept+for+Structure+Mooring

A moored ball will set it's moore point at 'range' from the surface of the relevant structure on the line towards the ship as found when the command is executed and will attempt to keep within ship.radius meters of that point during normal movement evolve.  

The collision response here isn't particularly sophisticated; when collided with, we always simply remove all the acceleration components that point away from the structure generated by the collision, so the ship may be bumped sideways and towards the structure, but never away from it.  This probably needs more work to avoid edge cases and exploits, but this is written as a tech-spike/prototype so I'm letting that slide for now. 

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/eve/src/": change = 1521217]