Maximum time to enter warp https://jira.ccpgames.com/browse/TFO-3069
mEffectStamp on the ball is already being used to mark a ball with DSTBALL_WARP as aligning by setting it to -1.  Changed it so that any negative value is interpreted as aligning whilst in that mode and decrementing it in each tick we evolve the ball in such an alignment phase.  If we find the ball in that phase with the absolute value of more than MAX_ALIGN_TICKS we force start the proper warp.

Also factored the checks for warp states into ball functions for better readability.

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/eve/src/": change = 1538183]