Prevent adding of balls during Evolve() (#8)
https://ccpgames.atlassian.net/browse/PLAT-1111

So, here's what I think is going on.

AddBall got called while the evolve loop was running. Why is this fatal?
Because when the inEvolve flag is set to "true", the ball does not get
added to mBalls when InsertBallInBoxes is called. This means that there
is nothing that holds a reference to the IRoot of the object after
Unlock() gets called on the IBall interface at the end of the function
(i.e. Nothing is referencing the Python object, so the Unlock results in
the reference count dropping to zero and the RootRefLock getting cleaned
up, explaining the missing entries in the virtual method table). By the
time BlueWrapObjectForPython gets called, the ball has already been
cleaned up.

How can Python code be running during the Evolve loop? There is just one
way that makes sense, and I have managed to reproduce the crash by
writing Python code to test this theory. The Evolve loop sends out DoXXX
events such as DoCollision, DoRange etc., that get processed right away.
These methods get run while the Evolve() method is still running, so if
they try to add balls to the park, then we run into this issue. I'm not
sure which Python callback method would have triggered the issue in that
particular case, but I suspect it may have something to do with
behaviors, which listen to destiny mode changes through DoModeChange,
and emits a Signal MESSAGE_ON_BALL_MODE_CHANGED, which gets processed in
the same thread, meaning that we are still within the Evolve() method.
This could potentially lead to tasks that add balls getting run within
that context, leading to a crash.

So, how do we fix this? I don't think we should allow things messing
with the ballpark as an Evolve() is happening, since we need to make
sure that the Evolve() results in identical results across the server
and clients, we need to be protective here and just raise an error.