Expose number of Destiny updates per tick (#11)
* Expose number of Destiny updates per tick

https://ccpgames.atlassian.net/browse/PLAT-2746 (Client-side gameplay updates can be processed before Destiny updates)

When the updates get processed client side, Destiny updates need to get processed before Gameplay updates. In order to make sure this is the case, the server sends a boolean flag along with any gameplay updates in order to tell the client to wait for the corresponding Destiny updates.

Destiny updates get sent out to players in two chunks:
- The first chunk gets sent out to everyone in the bubble
- The second chunk is just for the character who receives it (i.e. the user is in a cloaked ship)

We currently don't have enough information to know if we have received all the destiny updates for the current tick unless we receive both of these chunks or just the bubble one, which has an additional flag which tells us if we should wait for the character-only tick.

This change adds functionality so that each destiny tick includes the number of destiny tick batches that get sent to that character for that tick, so that the client can tell when the Destiny update has finished. It also adds methods for the game server code to check if a character or bubble has any pending Destiny updates, so that the Gameplay updates can include the wait-for-destiny flag when appropriate.