Integrate Orientation and Orbit changes from Frontier (#14)
https://ccpgames.atlassian.net/browse/PLAT-2770 (Integrate orientation changes into ccpgames/carbon-destiny:main)
https://ccpgames.atlassian.net/browse/PLAT-2771 (Integrate orbit changes into ccpgames/carbon-destiny:main)

Orientation and orbit can be enabled using Destiny settings.

GotoThrust has some changes in the Frontier branch. To avoid changing default behavior, I'm creating OrbitThrust which evolves orbit the same way that GotoThrust works on orbiting balls in the Frontier branch.

Regarding test_formation_follow_stopped_ball, expected coordinates changed because we are using the direction towards mGoto as the direction vector in CalculateYawPitchRoll, rather than using mGoto as the actual direction vector. (The formation cold is old abandoned code that needs to get removed or seriously worked on, because it relies on orientation which is not guaranteed to be equal between server and client because ball orientation gets "snapped" when a new ball is added to the park, which can result in inconsistencies between client/server when it comes to rotation for that tick, at least when using the old collision method. This formation code is not used for the formation flight which is currently available in Eve Online, that formation flight behavior was put together in Python)