Fix cryptic Python error when initializing bubbles for empty ballpark
In InitializeBubbles() after calling ClearBubbles(), the bubbles dict was set to nullptr. If there are any balls in the park, the dict would then get re-created when AddToBubble gets called right after. However, when there aren't any balls `bubbles` remained a nullptr.

In the following call to PyAdditionsAndDeletions, this would lead to returning a nullptr without setting the Python exception state, so we ended up with a misleading error message based on a previously run test. Therefore, change the ClearAll method to clear the bubbles dict instead of resetting it to a nullptr. Also, move into it the clearing of other bubble tracking python dicts/sets that always get cleared alongside the bubbles dict, whenever it is called.

Set the error state when calling AdditionsAndDeletions with an uninitialized bubbles dict, in order to avoid future confusion.

Remove redundant guard against null checks after PyArg_ParseTuple succeeds.